Add nvim-treesitter as a dependency, rearrange build deps
By using nvim-treesitter's queries, this plugin should be able to
progress substantially faster than if everything had to be written from
scratch. This also means that when the list of parsers has been set up
(and everything else works, obviously :p), it should work for all the
languages supported by nvim-treesitter.

Writing custom queries can be revisited when I have capacity for it, but
that time isn't now.

The new plan is: we include tree-sitter as a conan dependency, along
with IXWebSocket (a fantastic socket library that includes a server
socket. I've admittedly never used the server features, but there's a
first time for everything). IXWS tanks the problem of communicating with
Vim, as this isn't particularly easy from a language where most socket
libraries are written with Boost as a dependency.

This means that tree-sitter updates are dependent on the server, which
is actually pretty :blobthinksmart: because it prevents accidentally
breaking the plugin if tree-sitter has an update that has an
incompatible API change or something.
This largely means the server being recompiled is tree-sitter being
recompiled.
Using servers also means we can use one server for all instances of Vim